Rework structure, add log statistics classes
