+++ Analyzer/Program.cs
@@ -1,23 +1,32 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using Analyzer.Log;
 
 namespace Analyzer
 {
     class Program
     {
+        private static readonly FileGetter LogGetter = new FileGetter();
+
         static void Main(string[] args)
         {
-            FileGetter getter = new FileGetter();
-            List<FileInfo> logs = getter.GetLogs();
+            try
+            {
+                LogGetter.ImportLogs();
+            }
+            catch (ImportLogsException e)
+            {
+                Console.WriteLine(e.Message);
+                return;
+            }
+
+            List<FileInfo> logs = LogGetter.Logs;
             
             LogAnalyzer logAnalyzer = new LogAnalyzer();
             logAnalyzer.Start(logs);
-
-            Console.WriteLine();
-            Console.WriteLine("==============================================");
-            Console.WriteLine("Для завершения работы нажмите любую клавишу...");
-            Console.ReadKey(true);
+            
+            CliFormatter.PrintExitMessage();
         }
     }
 }
